+++ CHANGELOG.md
@@ -23,3 +23,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add tags to logger (request Id, IP address, User agent)
 - Separate Log files per environment
 - Implement `Securial.configuration` method with underlying classes
+- Implement configuration validation for logger settings